Proyecto Integración/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
python-decouple==3.8
Pillow==10.4.0
django-filter==23.3
setuptools>=68.0.0
